+++ src/public/virtual-keyboard.ts
@@ -267,12 +267,6 @@ export interface VirtualKeyboardInterface extends VirtualKeyboardOptions {
 // Commands return true if they resulted in a dirty state
 // @revisit: maybe a command attribute instead?
 export interface VirtualKeyboardCommands {
-  /**
-   * The command invoked when a variant key is pressed:
-   * hide the variants panel, then perform the command.
-   */
-  performVariant: (command: Selector | [Selector, ...any[]]) => boolean;
-
   switchKeyboardLayer: (layer: string) => boolean;
 
   toggleVirtualKeyboard: () => boolean;